requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
google-play-scraper==1.2.7
idna==3.10
macholib==1.16.3
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
requests==2.32.3
rumps==0.4.0
setuptools==75.3.0
urllib3==2.2.3
